Parcel Loader changes:  Kinds that don't have any superKinds assigned are automatically assigned //Schema/Core/Item as a superKind; Local attribute items are automatically linked to their parent kind item.
PyCrust:  A new menu item has been added to the "Test" menu which
will bring up the PyCrust debugger window; this includes an interactive
python shell, and also a data browser.  Expand "locals()" in the
lower-left corner to explore.  I have populated "locals()" with a
few handy starting points: "chandler" is the wxApplication, "Globals"
is application.Globals, "repo" is the repository, "pm" is the parcel
manager.