Adding experimental parcel reloading support; existing references to (and from) reloaded items are kept intact -- which means you can add references but not remove them (yet).  Literal values are cleared out and reloaded from XML.  You can also change the kind of an item.  Items may not be deleted (yet).
This required some changes to the notification manager which I need to go over
with Stuart.  The side-effect of reloading items is that their onItemLoad()
callback can get called before the notification manager's declarations list
has been populated.  To overcome this, I put in an additional call to
PrepareSubscribers( ) *before* parcel loading (in addition to the one
after parcel loading).  Also, I added some <initialValue> tags to
ItemCollection's attributes so that we don't need to initialize them in
ItemCollection's __init__ -- since that doesn't get called when items
get reloaded.