Bug 3250, bogus comparison prevented import of certain root certificates.
Bug 3220, use query subscriptions so that when you want to trust dynamically
added certificate permanently it actually works. Also had to move the
query to main thread, because the query on the mail thread was not seeing
changes made in the main thread - this is apparently a known issue.
This is also partial fix to adding certs dynamically - bug 3055.

Bug 3099, prevent importing the same certificate multiple times by checking
prior to import that it does not already exist.