First pass at the Preview Area by the minical - PLUS, crazy weird ?CPIA bug!
so there's now a preview area right above the minical (see calendar spec).
It's actually capable of displaying all the events of the currently selected
day.  It still has to learn to disappear when the calendar view is selected,
and be cleaned up a lot, but it's a start.

BUT there's this WEIRD bug: if you set the "contents" attribute, then the
SidebarContainer's layout gets REVERSED!  That's right, it's suddenly upside
down, with the minical&preview on top, and the sidebar proper on the bottom.

So I turned off the "contents" being set, but it needs contents to be set to
work, so the preview pane is non-functional for now...

Commenting/uncommentin the <contents/> line in main/parcel.xml will make the
preview pane work at the cost of making the sidebar layout upside down.  try it
for a fun time.  (search for "gui layout bug")