Resolve bug 3243 (Style probably should not be a ContentItem), by making Style a schema.Item.
When we tried this before, there was at least one biref that led to bug 3242 (calendar and detail views are blank with calendar filter), but I can't reproduce that now (possibly because of the schema changes). Unit tests pass, and I've seen no problems pounding on the app (mostly on PC, but also a little on Mac).

I'm checking in this change by itself in case it causes a problem - it can be rolled back on its own.