SplitterWindow better wx/cpia integration
It's now easy to move the sash of a SplitterWindow and get the position
properly saved at the CPIA/repo level.  In particular, the position moved to
via toggle button of the all-day area gets saved on resizes and switches away
and back to the calendar view.

This commit also adds lots of debug clutter, will remove soon I promise