Workaround for bug 3411. Fixes behavior on MacOS where Chandler starts with a height of one Toolbar-height too high. Uses wxFrame.SetSize() to resize the window to the (correctly) saved size.
NOTE: wxFrame.SetSize *should* (but may not) work as expected on every platform
and I have only tested on Mac. (please see that it works on other platforms).

NOTE: According to the wx documentation, wxFrame.SetClientSize() should be the
method of choice but that has even stranger behavior than the bug this is
trying to fix. (this is possibly also related to the toolbar) SEE bug 3682.